backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
groq
chromadb
pymupdf4llm
langchain
langchain-community
sentence-transformers
fpdf2
python-docx
pdfplumber
pytesseract
pdf2image
Pillow
httpx
